Fix SetupMatrix

+++ main.cpp
@@ -99,7 +99,7 @@ void CreateTopology(int systemsize, Epetra_SerialDenseMatrix& topology,
 void SetUpMatrix(Epetra_SerialDenseMatrix& A, std::vector<double> xv0,
                  std::vector<double> yv0, double delta, double E,
                  int systemsize, int k) {
-  int r;
+  double r;
   double pi = atan(1) * 4;
   double raggio = delta / 2;
   double C = 1 / (E * pi * raggio);
